feat(src): added superclass to project
