+++ src/filter.js
@@ -1,9 +1,9 @@
-import {createElement} from "./constants";
+import {Component} from "./component";
 
-export class Filter {
+export class Filter extends Component {
   constructor(data) {
+    super();
     this._filters = data;
-    this._element = null;
     this._filterClickHandler = this._filterClickHandler.bind(this);
   }
 
@@ -30,16 +30,5 @@ export class Filter {
 
   unbind() {
     this._element.removeEventListener(`click`, this._filterClickHandler);
-  }
-
-  render() {
-    this._element = createElement(this.template);
-    this.bind();
-    return this._element;
-  }
-
-  unrender() {
-    this.unbind();
-    this._element = null;
   }
 }
